optimisation: don't store the Response in static memory
Previously I thought declaring the `resp` variable as static
would allow it to be manipulated using direct addressing rather
than stack-relative addressive, which may be cheaper. However,
removing this specification actually reduces the size of the
compiled program. Perhaps the compiler is able to store
`resp` in registers rather than in RAM; in this instance I
can't determine the behaviour from the assembly dump.

Before:
   text    data     bss     dec     hex filename
   5462      86     600    6148    1804 build-uno/servo-asm-fw.elf

After:
   text	   data	    bss	    dec	    hex	filename
   5452	     86	    558	   6096	   17d0	build-uno/servo-asm-fw.elf