optimisation: enum sizes
Ensures all defined enum types specify the size of the integer
type that should be used to store them. Relevant because the
default is a 16-bit integer, but the two enums that are modified
by this commit have far fewer than 256 members and so could be
stored in an 8-bit integer.

Saves 20 bytes of compiled program space and a whopping 1 byte
of RAM usage.