Use PGM_P instead of const char* for strings in program memory
PGM_P [1] is the canonical type to use for a pointer to a string
that resides in program memory rather than in data memory (although
currently it is just an alias for `const char *`).

On a related note, PGM_VOID_P can be used for program memory
objects that are not strings.

[1]: http://www.nongnu.org/avr-libc/user-manual/group__avr__pgmspace.html#ga963f816fc88a5d8479c285ed4c630229